Reuse base handler code
This changes the structure to have a `DailySolution` base class with
methods to solve each part of the daily challenge. A base handler is
created that handles the JSON serialization and ensuring a body is
present. Each day then creates an implementation of `DailySolution` and
ideally exports it with the name `Solver` (to support the tests) and
passes it to `handlerBase` to automatically generate a handler for the
Function.

A future improvement would be to move the validation and error handling
into additional `middy` middleware and to potentially do the same even
for the `Solver` itself.